moved details tile to common
